#57: 텍스트 시작지점 상수 타입 수정
- 타입이 잘못 선언되어 있어 Int 타입으로 수정하였습니다.

+++ ToDo-Garden/ToDoGardenUI/Sources/ToDoGardenUIConstant/Constant+AddUnderlinedTextButton.swift
@@ -23,5 +23,5 @@ extension Constant.AddUnderlinedTextButton.Alpha {
 
 extension Constant.AddUnderlinedTextButton.Title {
   public static let baselineOffset: CGFloat = 5.0
-  public static let startPoint: CGFloat = 0
+  public static let startPoint: Int = 0
 }
